Move `_helpers` and `_constants` to `libs` directory; update imports accordingly in e2e tests
